fixes: call of overloaded 'IPAddress(unsigned int)' is ambiguous

+++ src/AsyncWebSocket.cpp
@@ -737,7 +737,7 @@ void AsyncWebSocketClient::binary(const __FlashStringHelper *data, size_t len)
 IPAddress AsyncWebSocketClient::remoteIP() const
 {
     if (!_client)
-        return IPAddress(0U);
+        return IPAddress((uint32_t)0);
 
     return _client->remoteIP();
 }
